extract overloading logic to base
